feat: Add lazy loading for images and update dependencies
This commit adds lazy loading functionality for images to improve page load performance. It also updates the npm dependencies to their latest stable versions.

The code changes include modifying the next.config.js file to include the \transpilePackages\ option for lazy loading the \three\ package. The tsconfig.json file is updated to include the \baseUrl\ and \paths\ options for resolving module imports. The .eslintrc.cjs file is modified to include the \@typescript-eslint/ban-ts-comment\ rule. The navbar.tsx component is updated to modify the border color. The flip-card.tsx component is updated to modify the border color. The package.json file is updated to add the \@react-three/drei\ and \@react-three/fiber\ packages as dependencies, and the \@types/three\ package as a dev dependency. The Loader.tsx component and Laptop.tsx component are added to the codebase.

These changes enhance the application's performance, maintainability, and user experience.